pyproject.toml
[project]
name = "tts-warehouse"
version = "0.1.0"
description = "Self-hosted TTS feed generator for articles, RSS feeds, and Royal Road books"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109",
    "uvicorn[standard]>=0.27",
    "sqlalchemy>=2.0",
    "pydantic-settings>=2.0",
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "feedparser>=6.0",
    "newspaper3k>=0.2",
    "python-dotenv>=1.0",
    "aiofiles>=23.0",
    "python-multipart>=0.0.22",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.2",
]

[tool.ruff]
line-length = 100
target-version = "py310"
